packages/backend/requirements.txt
# FastAPI backend package (packages/backend)
# 설치:
#   pip install -r requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
websockets>=12.0
requests>=2.31.0

# 데이터/유틸
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
psutil>=5.9.0

# ML (models/*.pt inference)
torch>=2.2.0

# GCS
google-cloud-storage>=3.4.0
google-auth>=2.0.0
